+++ extensions/converter/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## [Initial Version] - 2022-05-12
 
-- Convert number from any base to any base or convert between different character encodings.
+- Arbitrary input, direct output. Includes Base converter, Case converter, Code converter, etc.